+++ src/io/staticweb/cloudformation_templating.clj
@@ -17,16 +17,40 @@
 (defn arn [ref]
   {"Fn::GetAtt" [(full-name ref) "Arn"]})
 
-(defn base64 [x]
+(defn base64
+  "The intrinsic function Fn::Base64 returns the Base64 representation
+   of the input string. This function is typically used to pass encoded
+   data to Amazon EC2 instances by way of the UserData property.
+   
+   See https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/intrinsic-function-reference-base64.html"
+  [x]
   {"Fn::Base64" x})
 
-(defn cidr [ip-block count cidr-bits]
+(defn cidr
+  "The intrinsic function Fn::Cidr returns an array of CIDR address
+   blocks. The number of CIDR blocks returned is dependent on the count
+   parameter.
+   
+   See https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/intrinsic-function-reference-cidr.html"
+  [ip-block count cidr-bits]
   {"Fn::Cidr" [ip-block count cidr-bits]})
 
-(defn equals [x y]
+(defn equals
+  "Compares if two values are equal.
+   Returns true if the two values are equal or false if they aren't.
+   
+   See https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/intrinsic-function-reference-conditions.html#intrinsic-function-reference-conditions-equals"
+  [x y]
   {"Fn::Equals" [x y]})
 
 (defn find-in-map
+  "The intrinsic function Fn::FindInMap returns the value corresponding
+   to keys in a two-level map that's declared in the Mappings section.
+   
+   The 4-arity version with `default-value` requires the
+   AWS::LanguageExtensions transform.
+   
+   See https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/intrinsic-function-reference-findinmap.html"
   ([map-name top-level-key second-level-key]
    {"Fn::FindInMap"
     [(full-name map-name)
@@ -36,16 +60,46 @@
    (-> (find-in-map map-name top-level-key second-level-key)
        (update "Fn::FindInMap" conj {"DefaultValue" default-value}))))
 
-(defn fn-and [& conds]
+(defn fn-and
+  "Returns true if all the specified conditions evaluate to true,
+   or returns false if any one of the conditions evaluates to false.
+   Fn::And acts as an AND operator. The minimum number of conditions
+   that you can include is 2, and the maximum is 10.
+   
+   See https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/intrinsic-function-reference-conditions.html#intrinsic-function-reference-conditions-and"
+  [& conds]
   {"Fn::And" (vec conds)})
 
-(defn fn-if [cond then else]
+(defn fn-if
+  "Returns one value if the specified condition evaluates to true and
+   another value if the specified condition evaluates to false.
+   Currently, CloudFormation supports the Fn::If intrinsic function
+   in the metadata attribute, update policy attribute, and property
+   values in the Resources section and Outputs sections of a template.
+   You can use the [[no-value]] pseudo parameter as a return value to
+   remove the corresponding property.
+   
+   See https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/intrinsic-function-reference-conditions.html#intrinsic-function-reference-conditions-and"
+  [cond then else]
   {"Fn::If" [(full-name cond) then else]})
 
-(defn fn-not [cond]
+(defn fn-not
+  "Returns true for a condition that evaluates to false or returns
+   false for a condition that evaluates to true. Fn::Not acts as a
+   NOT operator.
+   
+   See https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/intrinsic-function-reference-conditions.html#intrinsic-function-reference-conditions-not"
+  [cond]
   {"Fn::Not" [cond]})
 
-(defn fn-or [& conds]
+(defn fn-or
+  "Returns true if any one of the specified conditions evaluate to true,
+   or returns false if all the conditions evaluates to false.
+   Fn::Or acts as an OR operator. The minimum number of conditions that
+   you can include is 2, and the maximum is 10.
+   
+   See https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/intrinsic-function-reference-conditions.html#intrinsic-function-reference-conditions-or"
+  [& conds]
   {"Fn::Or" (vec conds)})
 
 (defn for-each
@@ -77,13 +131,31 @@
   {(str "Fn::ForEach::" (full-name loop-name))
    [(full-name id) coll {(full-name output-key) output-val}]})
 
-(defn get-att [ref attr]
+(defn get-att
+  "The Fn::GetAtt intrinsic function returns the value of an attribute
+   from a resource in the template. When the AWS::LanguageExtensions
+   transform transform is used, you can use intrinsic functions as
+   a parameter to Fn::GetAtt.
+   
+   See https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/intrinsic-function-reference-getatt.html"
+  [ref attr]
   {"Fn::GetAtt" [(full-name ref) (full-name attr)]})
 
-(defn get-azs [& [region]]
+(defn get-azs
+  "The intrinsic function Fn::GetAZs returns an array that lists
+   Availability Zones for a specified Region in alphabetical order.
+   
+   See https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/intrinsic-function-reference-getavailabilityzones.html"
+  [& [region]]
   {"Fn::GetAZs" (or (full-name region) "")})
 
-(defn import-value [name]
+(defn import-value
+  "The intrinsic function Fn::ImportValue returns the value of an output
+   exported by another stack. You typically use this function to create
+   cross-stack references.
+   
+   See https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/intrinsic-function-reference-importvalue.html"
+  [name]
   {"Fn::ImportValue" (full-name name)})
 
 ; Logic inferred from
@@ -97,7 +169,14 @@
           (or (str/ends-with? ns ":")
             (str/index-of ns "::")))))))
 
-(defn join [separator coll]
+(defn join
+  "The intrinsic function Fn::Join appends a set of values into
+   a single value, separated by the specified delimiter. If a delimiter
+   is the empty string, the set of values are concatenated with no
+   delimiter.
+   
+   See https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/intrinsic-function-reference-join.html"
+  [separator coll]
   {"Fn::Join" [separator coll]})
 
 (defn length
@@ -110,9 +189,17 @@
   [array]
   {"Fn::Length" array})
 
-(def no-value {"Ref" "AWS::NoValue"})
+(def
+  ^{:doc
+    "Removes the corresponding resource property when specified
+     as a return value in the [[fn-if]] intrinsic function.
+     
+     See https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/pseudo-parameter-reference.html"}
+  no-value {"Ref" "AWS::NoValue"})
 
-(defn not-equals [x y]
+(defn not-equals
+  "The composition of [[fn-not]] and [[equals]]."
+  [x y]
   (fn-not (equals x y)))
 
 (defn outputs
@@ -149,25 +236,67 @@
       m
       m)))
 
-(defn ref [name]
+(defn ref
+  "The intrinsic function Ref returns the value of the specified
+   parameter or resource. When the AWS::LanguageExtensions transform
+   is used, you can use intrinsic functions as a parameter to Ref.
+   
+   See https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/intrinsic-function-reference-ref.html"
+  [name]
   {"Ref" (full-name name)})
 
-(def region
+(def
+  ^{:doc
+    "Returns a string representing the Region in which the encompassing
+     resource is being created, such as us-west-2.
+     
+     See https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/pseudo-parameter-reference.html#cfn-pseudo-param-region"}
+  region
   {:Ref "AWS::Region"})
 
-(defn select [index objects]
+(defn select
+  "The intrinsic function Fn::Select returns a single object from a
+   list of objects by index.
+   
+   See https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/intrinsic-function-reference-select.html"
+  [index objects]
   {"Fn::Select" [index objects]})
 
-(defn split [separator s]
+(defn split
+  "To split a string into a list of string values so that you can select
+   an element from the resulting string list, use the Fn::Split intrinsic
+   function. Specify the location of splits with a delimiter, 
+   such as , (a comma). After you split a string, use the Fn::Select
+   function to pick a specific element.
+   
+   See https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/intrinsic-function-reference-split.html"
+  [separator s]
   {"Fn::Split" [separator s]})
 
-(def stack-id
+(def
+  ^{:doc
+    "Returns the ID of the stack as specified with the
+    `aws cloudformation create-stack` command.
+     
+     See https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/pseudo-parameter-reference.html#cfn-pseudo-param-stackid"}
+  stack-id
   {:Ref "AWS::StackId"})
 
-(def stack-name
+(def
+  ^{:doc
+    "Returns the name of the stack as specified with the
+    `aws cloudformation create-stack` command.
+     
+     See https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/pseudo-parameter-reference.html#cfn-pseudo-param-stackname"}
+  stack-name
   {:Ref "AWS::StackName"})
 
-(defn sub [s]
+(defn sub
+  "The intrinsic function Fn::Sub substitutes variables in an
+   input string with values that you specify.
+   
+   See https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/intrinsic-function-reference-sub.html"
+  [s]
   {"Fn::Sub" s})
 
 (defn tags [& {:as m}]
@@ -191,7 +320,12 @@
   [object-or-array]
   {"Fn::ToJsonString" object-or-array})
 
-(defn transform [name parameters]
+(defn transform
+  "The intrinsic function Fn::Transform specifies a macro to perform
+   custom processing on part of a stack template.
+   
+   See https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/intrinsic-function-reference-transform.html"
+  [name parameters]
   {"Fn::Transform" {:Name name :Parameters parameters}})
 
 (defn user-data [& data]